pyproject.toml
[tool.poetry]
name = "order-assistant-webservice"
version = "0.1.0"
description = ""
authors = ["nadle"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "2.5.0"
django-environ = "0.10.0"
openai = "0.27.7"
django-bootstrap5 = "23.1"
requests = "*"
beautifulsoup4 = "*"
channels = "4.0.0"
gTTS = "2.3.2"
langchain = "*"
daphne = "^4.0.0"
django-widget-tweaks = "^1.5.0"
python-dotenv = "^1.0.0"
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

